refactor: extract GuardrailResolver from container.ts (429 → 317 lines)
Extract guardrail resolution functions (resolveGuardrails, resolveExecutionRails,
resolveRetrievalRails, listAvailableGuardrails, getEnforcedPluginIds, corePluginIds)
into a dedicated GuardrailResolver module.

container.ts remains above 150 lines (317) because the remaining content is
imports and composition root wiring, which is its core purpose.
